add getAllProjects Controller

+++ watchlog-web/src/main/java/com/netease/qa/log/web/controller/WlogSrcController.java
@@ -6,7 +6,6 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
-import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.bind.annotation.RequestParam;
@@ -55,9 +54,11 @@ public class WlogSrcController {
 			return new ResponseEntity<JSONObject>(apiException.handleNotFoundRequestException(nr), HttpStatus.NOT_FOUND);
 		}
 		int recordsTotal = logSourceService.getTotalCountByProjectId(Integer.parseInt(projectid));
-		String message = "select successful";
+		String message = Const.RESPONSE_SUCCESSFUL;
 		JSONArray data = logSourceService.getLogSourcesListByProjectid(Integer.parseInt(projectid), Integer.parseInt(limit),
 				Integer.parseInt(offset));
+		if(data == null)
+			message = Const.RESPONSE_NOTSUCCESSFUL;
 		JSONObject result = new JSONObject();
 		result.put("message", message);
 		result.put("recordsTotal", recordsTotal);
@@ -90,11 +91,11 @@ public class WlogSrcController {
 		}
 		String field = MathUtil.getSortField(sort);
 		int recordsTotal = logSourceService.getTotalCountByProjectId(Integer.parseInt(projectid));
-		String message = "select successful";
+		String message = Const.RESPONSE_SUCCESSFUL;
 		JSONArray data = logSourceService.getLogSourcesListSortedByProjectid(Integer.parseInt(projectid), field, order, Integer.parseInt(limit), Integer.parseInt(offset));
 		JSONObject result = new JSONObject();
 		if(data == null)
-			message = "select not successfully";
+			message = Const.RESPONSE_NOTSUCCESSFUL;
 		result.put("message", message);
 		result.put("total", recordsTotal);
 		result.put("rows", data);
@@ -103,6 +104,10 @@ public class WlogSrcController {
 	
 	@RequestMapping(value = "/***", method = RequestMethod.DELETE)
 	public ResponseEntity<JSONObject> deleteLogSources(@RequestParam(value = "ids") String ids, Model model) {
+		if(MathUtil.isEmpty(ids)){
+			NullParamException ne = new NullParamException(Const.NULL_PARAM);
+			return new ResponseEntity<JSONObject>(apiException.handleNullParamException(ne), HttpStatus.BAD_REQUEST);
+		}
 		int[] logsource_ids = MathUtil.parse2IntArray(ids);
 		int result = logSourceService.deleteLogSources(logsource_ids);
 		if (result == 0) {
